refactor: ClientApiGenTest split to multiple smaller
